+++ packages/veui/src/locale/zh-Hans/Transfer.js
@@ -6,10 +6,11 @@ i18n.register(
     available: '待选项',
     selected: '已选项',
     selectAll: '选择全部',
-    deselectAll: '取消选择全部',
+    deselectAll: '移除全部',
     expand: '展开项目',
     collapse: '收起项目',
-    noData: '没有待选项'
+    noData: '没有待选项',
+    select: '请选择'
   },
   {
     ns: 'transfer'