+++ packages/veui/src/mixins/i18n.js
@@ -4,7 +4,13 @@ import i18n from '../managers/i18n'
 export default {
   methods: {
     t (token, data) {
-      let key = getConfigKey(this.$options.name)
-      return i18n.get(key ? `${key}.${token}` : token, data)
+      let key = token
+      let [match, rootKey] = token.match(/^@([^@]+)$/) || []
+      if (match) {
+        key = rootKey
+      } else {
+        key = `${getConfigKey(this.$options.name)}.${token}`
+      }
+      return i18n.get(key, data)
     }
   }}